refactor: changing BuildEmailFormExtensible to BulkEmailForm
