+++ src/components/bulk-email-tool/bulk-email-form/BulkEmailForm/context/reducer.js
